Nose config to spaw logs when manual tests exec
When running some nose tests manually, we want to have debug logs available
on the shell. This config enables that by disabling nose capture (so even
passing tests will show the logs), show the test names for what we executed
and put debug mode.

We can run one test with:
$ nosetests3 -c log-confs/debug_test.cfg tests.small.test_download_center:TestDownloadCenter.test_download