Framework module loader
Inject framework system loader (with the android category and modules
scafolded).
A call to frameworks.load_frameworks() init this loading and populate
categories attach frameworks to them.
Add patchelem() context manager to be able to patch module attributes directly.
Add and shape a framework to test those with a dedicated test directory.
More tests to come in a following commit.