Add a way to return immediately to mainloop.
Fix a bug discovered by test when some warnings or errors can happen twice
when we recall ourself with sudo. We need to exit to the mainloop in the
parent process to not redo the same tests.
Knowing that setup can and is inherited, a simple return isn't enough, we raise
then a specific exception that is ignored by the mainloop, then the normal exit
cleanup can happen.