Add local repository creation facility
This local repository is stored inside the docker container, we can have
same arch or multi-arch components. It's picking the corresponding version at
the image creation time to ensure we can install them inside the container and
offline.

Use a subdirectory for the android-studio dependencies and add the needed
packages there.