Fix when a framework doesn't need a checksum
A recent change in master introduced Checksum() object in DownloadCenter.
The API supposed to pass None instead of this object if there is no checksum,
but BaseInstaller doesn't, it creates an object with None for each attribute
which makes sense. Add the support for this case to DownloadCenter with an
unit test.