Ensure we continue parsing the page after finding a download
We can have the license being after the download links in the page, so don't
break as soon as we have found the download page for that reason.

Tests will help ensuring we don't have any frameworks matching an inappropriate
tag later in the page. If so, we will make it more specific.