Add multi-hosts capability to our local server for medium tests
This enables to only set the base directory for the server content,
and dynamically translate the path relative to <base_dir>/<requested hostname>
for each requests.
That way, the local server can serve multiple hostnames on the same port, which
is needed for some frameworks making request on multiple domains, like dart.

Note that the ssl support for multiple hostnames (SNI) will be in an upcoming
commit.