Rename MAKE_IN_CONTAINER definition to be more explicit about what it is
This variable is used for tools helpers running in containers. However, it's
not about ensuring we are running this code for umake itself.
System tests are going to use fakely installed system packages and not that
local bindmount. However, we are still going to use that directory for running
tools helpers, like local server, checking if files exist and so on…