Enable running system umake as medium tests
Create a /usr/bin/umake file in the container and make it executable so that
we can execute "umake".
We bindmount the host package installation for the library. This way, starting
next release, we have access to system version (and not running from a
directory) to exercise real system running conditions.