Restore properly the os.environ object
Don't touch the original object, it seems that os.environ.copy() is returning
a dictionary copy of the object, not the original one. Consequently, when
you reassign this dict to os.environ, you loose the _Environ hidden object
which setenv() for you. Result is that subprocesses will then keep getting
the last environment set before first "os.environ =" restore.