Add locks to avoid a race condition when creating temp file
As installing new packages, updating the cache or adding new architectures will
require root permission, we ensure we don't create the temp download file in
case of a fast server answer with the same user account in a multi-threaded
context as it would have root permission then. This could particularly happens
in medium tests which has a local webserver. Adding locks to avoid this case.