Workaround gpg limitation refusing uid != euid.
When we install new packages, we are executing as root and then dropping
as the user for extracting and such. However, for signature verification,
we use gpg. This one doesn't like priviledge drop (if uid = 0 and
euid = 1000) and asserts if uid != euid.
Importing the key as root as well creates new gnupg files owned as root if
new keys weren't imported first.

Consequently, run gpg as root if we needed root access or as the user
otherwise. We store the gpg public key in a temporary gnupg directory that
will be removed under the same user rights (this directory needs to be owned
by the same user id to not be rejected by gpg).