Disable spring tools for now.
What happens is that it doesn't follow the baseinstaller framework process and
this has two side effects:
- There is some untrapped exceptions, which thus cascades in a stall process
- It now catches two urls, creating then two downloads (with the callback
called twice). The key is each time a different sha1sum, but only one is kept by
the base framework.

I think this is a good illustration why the baseinstaller should be used as
often as possible, as it has guards against such issues, ensure exceptions
are trapped and so on.
I think we should look for this framework to use back the main framework and
only catch latest (higher?) download link to have the callback only once.

For reference, the PR once it was merged was:
https://github.com/ubuntu/ubuntu-make/pull/240